pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-lab"
version = "0.1.0"
description = "Prompt Engineering Evaluation & Optimization System"
authors = [
    { name = "YOUR_NAME_HERE" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "python-dotenv",
    "pyyaml",
    "openai>=1.0.0",
]

[project.scripts]
prompt-lab = "prompt_lab.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]
